test_project/requirements.txt
pytest
pytest-django
splinter
django-extensions
django-bootstrap3
pep8
autopep8
-e ..